Updated Calculator with React Hooks
