+++ Learning/Gatsby/firstLook/src/components/calculatorKey.js
@@ -2,14 +2,14 @@ import React from "react";
 
 function CalculatorKey(props) {
     return (
-        <td>
+
             <input className={`${props.className}`}
                 type="button"
                 value={props.value}
                 onClick={() => props.onClick(props.value)}
             >
             </input>
-        </td>
+
     );
 }
