+++ Learning/Gatsby/firstLook/src/styles/styles.scss
@@ -336,6 +336,16 @@ header {
     }
 }
 
+// 404
+.fourofour {
+    padding: 100px 10px;
+    width: 100%;
+    position: relative;
+    align-items: center;
+    text-align: center;
+}
+
+
 //Calculator
 
 .calculator {
@@ -366,24 +376,27 @@ header {
         .body{
             background: $off-white;
             padding: 20px 20px 20px 25px;
-            width: 350px;
+            width: 320px;
             height: 570px;
             border-radius: 3px;
             border: 1px solid #c4c4c4;
             display: grid;
             grid-gap: 20px;
             .display{
+                margin: 10px 0;
                 border-radius: 3px;
                 border: 1px solid #c4c4c4;
                 background-color: #000000c9;
                 height: 50px;
-                width: 220px;
+                width: 260px;
                 font-size: 2vmax;
                 text-align: right;
                 color: $off-white;
                 padding-right: 10px;
+                align-items: center;
             }
             .btn{
+                margin: 0 10px;
                 height: 50px;
                 width: 50px;
                 border-radius: 3px;
@@ -394,6 +407,7 @@ header {
                 background-image: linear-gradient(to bottom,transparent,transparent 50%,rgba(0,0,0,.04));
                 box-shadow: inset 0 0 0 1px rgba(255,255,255,.05), inset 0 1px 0 0 rgba(255,255,255,.45), inset 0 -1px 0 0 rgba(255,255,255,.15), 0 1px 0 0 rgba(255,255,255,.15);
                 text-shadow: 0 1px rgba(255,255,255,.4);
+                position: relative;
                 &:hover{
                     background-color: #eaeaea;
                 }
@@ -409,6 +423,8 @@ header {
                     background-color: #2e86c0;
                     border-color: #337cac;
                     color: $off-white;
+                    // position: absolute;
+                    // height: 145px;
                     &:hover{
                         background-color: #4e9ed4;
                     }
@@ -416,6 +432,16 @@ header {
                 &.opr{
                     color: #4e9ed4;
                     font-size: 200%;
+                }
+                &.zero {
+                    // position: absolute;
+                    width: 120px;
+                }
+                &:first-child{
+                    margin-left: 0;
+                }
+                &:last-child{
+                    margin-right: 0;
                 }
             }
